requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dj-rest-auth==2.2.5
Django==4.2
django-cors-headers==3.14.0
django-crontab==0.7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
dymmond-settings==1.0.5
email-validator==2.2.0
exceptiongroup==1.2.2
idna==3.7
itsdangerous==2.2.0
jinja2==3.1.4
lilya==0.9.0
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgspec==0.18.6
multidict==6.0.5
nest-asyncio==1.6.0
orjson==3.10.6
Pillow==8.4.0
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-settings==2.3.4
pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
requests==2.32.3
rich==13.7.1
sniffio==1.3.1
sqlparse==0.5.1
text-unidecode==1.3
typing-extensions==4.12.2
urllib3==2.2.2
